Colocate CTAS unit tests
